+++ src/connections/CloneKernelConn.hpp
@@ -34,14 +34,14 @@ public:
 
    virtual int setParams(PVParams * params);
 
-   virtual int updateState(float time, float dt);
+   virtual int updateState(double time, double dt);
 
    // virtual int writeWeights(PVPatch *** patches, pvdata_t ** dataStart, int numPatches,
    //      const char * filename, float timef, bool last){return PV_SUCCESS;};
-   virtual int writeWeights(float time, bool last=false){return PV_SUCCESS;};
+   virtual int writeWeights(double time, bool last=false){return PV_SUCCESS;};
    virtual int writeWeights(const char * filename){return PV_SUCCESS;};
    virtual int checkpointWrite(const char * cpDir){return PV_SUCCESS;};
-   virtual int checkpointRead(const char * cpDir, float *timef){return PV_SUCCESS;};
+   virtual int checkpointRead(const char * cpDir, double *timef){return PV_SUCCESS;};
 
 protected:
    virtual PVPatch *** initializeWeights(PVPatch *** patches, pvdata_t ** dataStart, int numPatches,
